Refactor: extract helper functions
